+++ cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 ## Version 1
+### Version 1.1.0
+* Organizations that the -g user contributed to are now supported.
+* Better error-type-thing reporting.
+* Catches Octokit::Unauthorized errors and returns false when initializing the client.
+* Fix user_exists? error catching minor syntax bug.
+
+
 ### Version 1.0.2
 * Actually fix load path stuff again.